add back animations

+++ src/util/animation.cpp
@@ -8,7 +8,7 @@ UnitBezier ease(0.25, 0.1, 0.25, 1);
 
 animation::~animation() {}
 
-ease_animation::ease_animation(double from, double to, double &value, double duration)
+ease_animation::ease_animation(double from, double to, double &value, float duration)
     : animation(duration),
       from(from),
       to(to),
@@ -16,7 +16,7 @@ ease_animation::ease_animation(double from, double to, double &value, double dur
 }
 
 animation::state ease_animation::update() const {
-    double t = progress();
+    float t = progress();
     if (t >= 1) {
         value = to;
         return complete;
